rename config files to prevent linting errors
